==> code/0_tweet_extraction.ipynb <==
!pip install tweepy==4.12.1

import tweepy

tweepy.__version__

# Authenticate with your Twitter API keys and access tokens
bearer_token = 'AAAAAAAAAAAAAAAAAAAAAD5azgAAAAAAuG8CxzVp42RacKxMVN4%2FZ1hbQn4%3DK5RR8FfXnVTyVU7vGJVMZLyD20ZarLKnkv6Lk5bsGztvWqP7Lg'

# Create a Tweepy API object
auth = tweepy.auth.OAuth2BearerHandler(bearer_token)
api = tweepy.API(auth)

# Define the search query and the search period
query = "#diet"
start_date = "2022-12-01"
end_date = "2022-12-01"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
diet_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    diet_tweets.append(tweet)

len(diet_tweets)

# Define the search query and the search period
query = "#caloriedeficit"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
caloriedeficit_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    caloriedeficit_tweets.append(tweet)

len(caloriedeficit_tweets)

# Define the search query and the search period
query = "#workout"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
workout_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    workout_tweets.append(tweet)

len(workout_tweets)

# Define the search query and the search period
query = "#fitnessmotivation"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
fitnessmotivation_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    fitnessmotivation_tweets.append(tweet)

len(fitnessmotivation_tweets)

# Define the search query and the search period
query = "#lowcal"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
lowcal_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    lowcal_tweets.append(tweet)

len(lowcal_tweets)

# Define the search query and the search period
query = "#healthy"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
healthy_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    healthy_tweets.append(tweet)

len(healthy_tweets)

# Define the search query and the search period
query = "#proana"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
proana_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    proana_tweets.append(tweet)

len(proana_tweets)

# Define the search query and the search period
query = "#fitspo"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
fitspo_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    fitspo_tweets.append(tweet)

len(fitspo_tweets)

# Define the search query and the search period
query = "#hipbones"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
hipbones_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    hipbones_tweets.append(tweet)

len(hipbones_tweets)

# Define the search query and the search period
query = "#promia"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
promia_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    promia_tweets.append(tweet)

len(promia_tweets)

# Define the search query and the search period
query = "#anorexia"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
anorexia_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    anorexia_tweets.append(tweet)

len(anorexia_tweets)

# Define the search query and the search period
query = "#thinspiration"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
thinspiration_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    thinspiration_tweets.append(tweet)

len(thinspiration_tweets)

# Define the search query and the search period
query = "#41DaysofStarvation"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
DaysofStarvation_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    DaysofStarvation_tweets.append(tweet)

len(DaysofStarvation_tweets)

# Define the search query and the search period
query = "#bonespo"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
bonespo_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    bonespo_tweets.append(tweet)

len(bonespo_tweets)

# Define the search query and the search period
query = "#thinspo"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
thinspo_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    thinspo_tweets.append(tweet)

len(thinspo_tweets)

# Define the search query and the search period
query = "#weightloss"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
weightloss_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    weightloss_tweets.append(tweet)

len(weightloss_tweets)

# Define the search query and the search period
query = "#bulimia"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
bulimia_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    bulimia_tweets.append(tweet)

len(bulimia_tweets)

# Define the search query and the search period
query = "#edtwt"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
edtwt_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    anamia_tweets.append(tweet)

len(anamia_tweets)

# Define the search query and the search period
query = "#edtwt"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
edtwt_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    edtwt_tweets.append(tweet)

len(edtwt_tweets)

# Define the search query and the search period
query = "#meanspo"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
meanspo_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    meanspo_tweets.append(tweet)

len(meanspo_tweets)

# Define the search query and the search period
query = "#eatingdisorders"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
eatingdisorders_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    eatingdisorders_tweets.append(tweet)

len(eatingdisorders_tweets)

# Define the search query and the search period
query = "#dysmorphia"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
dysmorphia_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    dysmorphia_tweets.append(tweet)

len(dysmorphia_tweets)

# Define the search query and the search period
query = "#mentalhealth"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
mentalhealth_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    mentalhealth_tweets.append(tweet)

len(mentalhealth_tweets)

# Define the search query and the search period
query = "#EdTwitter"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
EdTwitter_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    EdTwitter_tweets.append(tweet)

len(EdTwitter_tweets)

# Define the search query and the search period
query = "#depression"

# Use the API's search premium endpoint to retrieve all tweets containing the hashtag within the specified period
depression_tweets = []
#for tweet in tweepy.Cursor(api.search_tweets, q=query, since=start_date, until=end_date).items():
for tweet in tweepy.Cursor(api.search_tweets, q=query).items():
    depression_tweets.append(tweet)

len(depression_tweets)

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

import json

for idx, tweet in enumerate(EdTwitter_tweets):
  with open('./drive/MyDrive/tweets/#EdTwitter_tweets/tweet_#EdTwitter_tweets_' + str(idx) + '.json', 'w') as f:
    json.dump(tweet._json, f, indent=4)
