==> code/1_dataframe_extraction.ipynb <==
## Mount al drive, instalación NetworkX e importaciones

import pandas as pd
import warnings
import itertools
import json
import os

warnings.simplefilter(action='ignore', category=FutureWarning)

!pip install -U networkx==2.8.4

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

## Funciones de extracción de nodos y aristas del tweet y creación de filas

def source_usernames_and_target_extractor( data : dict, source_name: list,  usernames : list, target_names: list) :
  # extract the usernames of every account that may appear in the tweet

  #extract the ursername of the main user
  main_user = data['user']['screen_name']
  source_name.append(main_user)
  usernames.append(main_user)

  if 'entities' in data:
      entities = data['entities']
    
      # extract usernames from user mentions
      if 'user_mentions' in entities:
          mentions = entities['user_mentions']
          for mention in mentions:
              username = mention['screen_name']
              usernames.append(username)
              target_names.append(username)
    
      # extract usernames from tweet text
      if 'full_text' in data:
          text = data['full_text']
      elif 'text' in data:
          text = data['text']
    
      words = text.split()
      for word in words:
          if word.startswith('@') and len(word[1:]) > 0:
              username = word[1:]
              if username[-1] != ':' :
                usernames.append(username)
                target_names.append(username)

  # extract usernames from retweeted_status
  if 'retweeted_status' in data:
      retweeted_data = data['retweeted_status']

      # extract username from user screen name
      if 'user' in retweeted_data and 'screen_name' in retweeted_data['user']:
          usernames.append(retweeted_data['user']['screen_name'])
          target_names.append(retweeted_data['user']['screen_name'])

      # extract usernames from entities
      if 'entities' in retweeted_data:
          entities = retweeted_data['entities']

          # extract usernames from user mentions
          if 'user_mentions' in entities:
              mentions = entities['user_mentions']
              for mention in mentions:
                  username = mention['screen_name']
                  usernames.append(username)
                  target_names.append(username)

          # extract usernames from tweet text
          if 'full_text' in retweeted_data:
              text = retweeted_data['full_text']
          elif 'text' in retweeted_data:
              text = retweeted_data['text']

          words = text.split()
          for word in words:
              if word.startswith('@') and len(word[1:]) > 0:
                  username = word[1:]
                  if username[-1] != ':' :
                    usernames.append(username)
                    target_names.append(username)


def create_row(df, target, source_name, hashtag):
  idx = ((df['source'] == target) & (df['target'] == source_name[0]))
  if not df.loc[idx].empty:
    row = {'source': target,
               'target': source_name[0],
               'weight': 1,
               'hashtags': hashtag}
  else:
    row = {'source': source_name[0],
               'target': target,
               'weight': 1,
               'hashtags': hashtag}
  return row

## Función para obtener dataframe a partir de json_objects

def get_dataframe_from_tweets(json_objects):
  # create an empty dataframe with the desired columns
  df = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])

  usernames = []
  # iterate over each JSON object
  for json_obj in json_objects:
      source_name = []
      target_names = []

      source_usernames_and_target_extractor( json_obj, source_name,  usernames, target_names)

      #print('source: ', source_name, '\n', 'targets: ',target_names, '\n')

      # add a row to the dataframe for each target
      for target in target_names:
          # create a row for the current source-target pair
          row = create_row(df, target, source_name, hashtag)

          # add the new row to the dataframe              
          df = df.append(row, ignore_index=True)
          #print(df, '\n')
    

  #print('usernames: ', set(usernames), '\n')

  # generate all combinations of tuples of strings in the set
  #combinations = list(itertools.combinations(set(usernames), 2))
 # for edge in combinations:
    #source_name = [edge[0]]
    #target = edge[1]
    #row = create_row(df, target, source_name, hashtag)
    #df = df.append(row, ignore_index=True)

  return df

## Dataframes extracción

# MENTAL HEALTH

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#shtwt_tweets', '#depression_tweets']
df_mental_health = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_mental_health = pd.concat([df_mental_health, df_aux])

# FITNESS

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#healthy_tweets','#fitspo_tweets', '#weightloss_tweets']
df_fitness = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_fitness = pd.concat([df_fitness, df_aux])

# EATING DISORDERS VERSION 1

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#ricecaketwt_tweets', '#proana_tweets','#thinspo_tweets', '#anorexia_tweets', '#eatingdisorder_tweets', '#anatwt_tweets', '#edtwtdiet_tweets', '#bulimia_tweets', '#meanspo_tweets', '#promia_tweets', '#anamia_tweets','#EdTwitter_tweets']
df_disorders_1 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_disorders_1 = pd.concat([df_disorders_1, df_aux])

# EATING DISORDERS VERSION 2

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#edtwt_tweets']
df_disorders_2 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_disorders_2 = pd.concat([df_disorders_2, df_aux])

# MENTAL HEALTH HASHTAG

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#mentalhealth_tweets']
df_mental_health_2 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_mental_health_2 = pd.concat([df_mental_health_2, df_aux])

# WORKOUT HASHTAG

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#workout_tweets']
df_fitness_2 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_fitness_2 = pd.concat([df_fitness_2, df_aux])

# FITNESS REMAINING HASHTAGS

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#diet_tweets','#caloriedeficit_tweets', '#fitnessmotivation_tweets', '#lowcal_tweets']
df_fitness_3 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_fitness_3 = pd.concat([df_fitness_3, df_aux])

# DISORDERS REMAINING HASHTAGS

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = ['#bonespo_tweets','#eatingdisorders_tweets', '#edtwtthread_tweets', '#3dtwt_tweets']
df_disorders_3 = pd.DataFrame(columns=['source', 'target', 'weight', 'hashtags'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects)
    print(hashtag , len(json_objects))
  df_disorders_3 = pd.concat([df_disorders_3, df_aux])

## Visualización de dataframes y modificaciones

df_mental_health

df_mental_health_weighted = df_mental_health.groupby(["source", "target"]).weight.sum().reset_index()
df_mental_health_weighted

df_fitness

df_fitness_weighted = df_fitness.groupby(["source", "target"]).weight.sum().reset_index()
df_fitness_weighted

df_disorders_1

df_disorders_1_weighted = df_disorders_1.groupby(["source", "target"]).weight.sum().reset_index()
df_disorders_1_weighted

df_disorders_2

df_disorders_2_weighted = df_disorders_2.groupby(["source", "target"]).weight.sum().reset_index()
df_disorders_2_weighted

df_mental_health_2

df_mental_health_2_weighted = df_mental_health_2.groupby(["source", "target"]).weight.sum().reset_index()
df_mental_health_2_weighted

df_fitness_2

df_fitness_2_weighted = df_fitness_2.groupby(["source", "target"]).weight.sum().reset_index()
df_fitness_2_weighted

df_fitness_3

df_fitness_3_weighted = df_fitness_3.groupby(["source", "target"]).weight.sum().reset_index()
df_fitness_3_weighted

df_disorders_3

df_disorders_3_weighted = df_disorders_3.groupby(["source", "target"]).weight.sum().reset_index()
df_disorders_3_weighted

## Guardar los dataframes

df_mental_health.to_csv('mental_health_graph_dataframe.csv', index=False)

df_fitness.to_csv('fitness_graph_dataframe.csv', index=False)

df_disorders_1.to_csv('disorders_1_graph_dataframe.csv', index=False)

df_disorders_2.to_csv('disorders_2_graph_dataframe.csv', index=False)

df_mental_health_2.to_csv('mental_health_2_graph_dataframe.csv', index=False)

df_fitness_2.to_csv('fitness_2_graph_dataframe.csv', index=False)

df_fitness_3.to_csv('fitness_3_graph_dataframe.csv', index=False)

df_disorders_3.to_csv('disorders_3_graph_dataframe.csv', index=False)

import networkx as nx
G = nx.from_pandas_edgelist(df_fitness_weighted, source="source", target="target", edge_attr="weight", create_using=nx.Graph())

from networkx.readwrite import gexf
gexf.write_gexf(G, "my_graph.gexf")
