==> code/2_dataframe_info_tweets.ipynb <==
## Mount al drive e importaciones

import pandas as pd
import warnings
import itertools
import json
import os

warnings.simplefilter(action='ignore', category=FutureWarning)

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

## Función para obtener dataframe a partir de json_objects

def get_dataframe_from_tweets(json_objects, hashtag):
  # create an empty dataframe with the desired columns
  df = pd.DataFrame(columns=['user', 'user_id', 'text','created_at', 'lang', 'hashtag', 'retweet_count', 'favorite_count'])


  # iterate over each JSON object
  for data in json_objects:
    user = data['user']['screen_name']
    user_id = data['user']['id_str']
    text = data['text']
    created_at = data['created_at']
    lang = data['lang']
    retweet_count = data['retweet_count']
    favorite_count = data['favorite_count']


    # add a row to the dataframe for each tweet
    row = {'user': user,
              'user_id': user_id,
              'text': text,
              'created_at' : created_at,
              'lang' : lang,
              'hashtag': hashtag,
              'retweet_count': retweet_count,
              'favorite_count' : favorite_count}

    # add the new row to the dataframe              
    df = df.append(row, ignore_index=True)

    
  return df

## Extracción de tweets en dataframes

path = '/content/drive/MyDrive/Computación/tweets/'
hashtags = os.listdir(path)
df_tweet_info = pd.DataFrame(columns=['user', 'user_id', 'text','created_at', 'lang', 'hashtag', 'retweet_count', 'favorite_count'])
for hashtag in hashtags:
  if hashtag.startswith('#'):
    tweets = os.listdir(path+hashtag)
    json_objects = []
    for tweet in tweets:
      with open(path + hashtag + '/' + tweet, 'r') as f:
        data = json.load(f)
        json_objects.append(data)
    
    df_aux = get_dataframe_from_tweets(json_objects, hashtag)
    print(hashtag , len(json_objects))
  df_tweet_info = pd.concat([df_tweet_info, df_aux])

## Visualización dataframe

df_tweet_info

## Guardar dataframe

df_tweet_info.to_csv('all_tweets_info_dataframe.csv', index=False)
