==> code/3_dataframe_cleaning.ipynb <==
## Mount al drive, instalación e importaciones

import pandas as pd
import itertools
import json
import os

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

!pip install -U networkx==2.8.4

import networkx as nx

## Load dataframes

path = '/content/drive/MyDrive/Computación/dataframes/'
dataframes = os.listdir(path)

dataframes

df_mental_health = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/mental_health_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_1 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_1_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_2_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_mental_health_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/mental_health_2_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_2_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness_3 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_3_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_3 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_3_graph_dataframe.csv", sep= ',',lineterminator='\n', index_col=None)

## Checking users 

import re
username_regex =  r'([\w\.]+)'

list_of_dataframes = [df_mental_health , df_fitness , df_disorders_1 , df_disorders_2 , df_mental_health_2 , df_fitness_2 , df_fitness_3 , df_disorders_3]
for dataframe in list_of_dataframes:
  dataframe['target'] = dataframe['target'].str.extract(username_regex, expand= True) #drop filas que tengan none después de extracción
  dataframe['source'] = dataframe['source'].str.extract(username_regex, expand=True) #podemos aplicar tranquilamente

for change in list_of_dataframes:
  if change['source'].isna().any():
      print('Al menos un elemento en la columna A es None')
  else:
      print('No hay elementos None en la columna A')

#df_mental_health_2, df_fitness_2,  df_fitness_2 drop where none after change in target
for change in list_of_dataframes:
  if change['target'].isna().any():
      print('Al menos un elemento en la columna A es None')
      indices_nulos = change.index[change['target'].isnull()].tolist()
      print(indices_nulos)
      change.drop(indices_nulos, inplace=True)

  else:
      print('No hay elementos None en la columna A')
#df_mental_health_2 , df_fitness_2 , df_fitness_3

for change in list_of_dataframes:
  if change['target'].isna().any():
      print('Al menos un elemento en la columna A es None')
      print(change.index[change['target'].isnull()].tolist())
  else:
      print('No hay elementos None en la columna A')
#df_mental_health_2 , df_fitness_2 , df_fitness_3

## Correcting hashtags

import re

list_of_dataframes = [df_mental_health , df_fitness , df_disorders_1 , df_disorders_2 , df_mental_health_2 , df_fitness_2 , df_fitness_3 , df_disorders_3]
for dataframe in list_of_dataframes:
  dataframe['hashtags'] = dataframe['hashtags'].apply(lambda x: re.findall(r'#\w+(?=_)', x)[0]) #podemos aplicar tranquilamente

for change in list_of_dataframes:
  if change['hashtags'].isna().any():
      print('Al menos un elemento en la columna A es None')
  else:
      print('No hay elementos None en la columna A')


## Guardar dataframes limpios

df_mental_health.to_csv('mental_health_graph_dataframe_final.csv', index=False)

df_fitness.to_csv('fitness_graph_dataframe_final.csv', index=False)

df_disorders_1.to_csv('disorders_1_graph_dataframe_final.csv', index=False)

df_disorders_2.to_csv('disorders_2_graph_dataframe_final.csv', index=False)

df_mental_health_2.to_csv('mental_health_2_graph_dataframe_final.csv', index=False)

df_fitness_2.to_csv('fitness_2_graph_dataframe_final.csv', index=False)

df_fitness_3.to_csv('fitness_3_graph_dataframe_final.csv', index=False)

df_disorders_3.to_csv('disorders_3_graph_dataframe_final.csv', index=False)

df_mental_health_weighted = df_mental_health.groupby(["source", "target"]).weight.sum().reset_index()
df_mental_health_weighted

G = nx.from_pandas_edgelist(df_fitness_weighted, source="source", target="target", edge_attr="weight", create_using=nx.Graph())

from networkx.readwrite import gexf
gexf.write_gexf(G, "my_graph.gexf")
