==> code/4_graphs_creation.ipynb <==
## Mount al drive, instalación e importaciones

import pandas as pd
import itertools
import json
import os

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

!pip install -U networkx==2.8.4

import networkx as nx

## Loading dataframes

path = '/content/drive/MyDrive/Computación/dataframes/'
dataframes = os.listdir(path)
print(dataframes)

df_mental_health = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/mental_health_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_1 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_1_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_2_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_mental_health_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/mental_health_2_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness_2 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_2_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_fitness_3 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/fitness_3_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

df_disorders_3 = pd.read_csv("/content/drive/MyDrive/Computación/dataframes/disorders_3_graph_dataframe_final.csv", sep= ',',lineterminator='\n', index_col=None)

## Combinaciones grafos

list_of_dataframes = [df_mental_health , df_fitness , df_disorders_1 , df_disorders_2 , df_mental_health_2 , df_fitness_2 , df_fitness_3 , df_disorders_3]

df_all_dataframes = pd.concat(list_of_dataframes, ignore_index=True)

df_all_dataframes

# agregar columna nueva
df_all_dataframes['initial_community'] = 0
disorders =['#ricecaketwt', '#proana', '#thinspo', '#anorexia','#eatingdisorder', '#anatwt','#edtwtdiet','#bulimia','#meanspo','#promia','#anamia','#EdTwitter','#edtwt','#bonespo','#eatingdisorders','#edtwtthread','#3dtwt']
health = ['#shtwt','#depression','#mentalhealth']
fitness = ['#healthy','#fitspo','#weightloss','#workout','#diet','#lowcal','#caloriedeficit','#fitnessmotivation']

# recorrer filas del DataFrame y agregar valores a la nueva columna según el valor de la columna "hashtags"
for i, row in df_all_dataframes.iterrows():
    if row['hashtags'] in disorders:
        df_all_dataframes.at[i, 'initial_community'] = 3
    if row['hashtags'] in health:
      df_all_dataframes.at[i, 'initial_community'] = 2
    if row['hashtags'] in fitness:
        df_all_dataframes.at[i, 'initial_community'] = 1
    #else:
      #df_all_dataframes.at[i, 'initial_community'] = 2

df_all_dataframes['initial_community'].unique()

df_all_dataframes_weighted = df_all_dataframes.groupby(["source", "target"]).weight.sum().reset_index()
df_all_dataframes_weighted

df_all_dataframes_weighted_hash = df_all_dataframes.groupby(["source", "target", 'initial_community']).weight.sum().reset_index()
df_all_dataframes_weighted_hash

from networkx.algorithms.assortativity.mixing import node_attribute_xy
all_communities_graph = nx.from_pandas_edgelist(df_all_dataframes_weighted, source="source", target="target", edge_attr="weight", create_using=nx.Graph())

all_communities_graph_hash = nx.from_pandas_edgelist(df_all_dataframes_weighted_hash, source="source", target="target", edge_attr=["weight",'initial_community'], create_using=nx.Graph())

## Guardar grafos formato gephi

from networkx.readwrite import gexf
gexf.write_gexf(all_communities_graph, "all_communities_graph.gexf")

gexf.write_gexf(all_communities_graph_hash, "all_communities_graph_hash.gexf")
